% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.se.R
\name{clustered.se}
\alias{clustered.se}
\title{This function allows for calculation of clustered standard errors after running a standard linear regression with lm.}
\usage{
clustered.se(model_result, data, cluster, level = 0.95, Tvar = "treatment")
}
\arguments{
\item{model_result}{Model object returned by lm}

\item{data}{The data set used in the original regression}

\item{cluster}{The name of the clustering variable}

\item{level}{The confidence level at which the p-value should be calculated}

\item{Tvar}{The name of the treatment variable from the original regression}
}
\value{
A list containing a variance-covariance matrix, clustered standard errors and p-values for each coefficient in the original model, and lower and upper bounds for confidence intervals at each specified confidence level.
}
\description{
This function allows for calculation of clustered standard errors after running a standard linear regression with lm.
}

