% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact}
\alias{impact}
\title{Title}
\usage{
impact(data = NULL, outcome_var = NULL, treat_var = NULL,
  control_vars = NULL, cluster_var = NULL, grade_var = NULL,
  direction = NULL, cutoff = NULL, probability = 75)
}
\arguments{
\item{data}{CSV data file input}

\item{outcome_var}{Character vector of length 1 indicating outcome variable}

\item{treat_var}{Character vector of length 1 indicating treatment variable}

\item{control_vars}{Character vector indicating control variables (optional)}

\item{cluster_var}{Character vector indicating a clustering variable (optional)}

\item{grade_var}{Character vector of length 1 indicating grade variable. If NULL, analysis will not split data by grade. Defaults to NULL.}

\item{direction}{The user-entered direction of intended impact.}

\item{cutoff}{The user-entered cutoff at which they would consider a program to have "moved the needle."}

\item{probability}{The user-entered minimum probability of the estimated impact being greater than the cutoff value to consider the program to have "moved the needle."}
}
\value{
A list containing an error message if applicable, and if no errors are found, impact results by grade (if grade_var is not specified, a single matching result is returned).
}
\description{
Title
}
\examples{

}
