% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{Title}
\usage{
matching(data = NULL, treat_var = NULL, match_vars = NULL,
  grade_var = NULL)
}
\arguments{
\item{data}{CSV data file input}

\item{treat_var}{Character vector of length 1 indicating treatment variable}

\item{match_vars}{Character vector of length at least 1 indicating matching variables}

\item{grade_var}{Character vector of length 1 indicating grade variable. If NULL, analysis will not split data by grade. Defaults to NULL.}
}
\value{
A list containing an error message if applicable, and if no errors are found, matching results by grade (if grade_var is not specified, a single matching result is returned).
}
\description{
Title
}
\examples{
match_by_grade <- matching(
data = matching_by_grade,
treat_var = 'Treatment',
match_vars = 'pre.test',
grade_var = 'grade')

}

