% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_by_grade.R
\docType{data}
\name{matching_by_grade}
\alias{matching_by_grade}
\title{Data.frame that includes randomly generated variables to be used in tests and examples.}
\description{
Data.frame that includes randomly generated variables to be used in tests and examples.
}
\keyword{data}
