% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{Title}
\usage{
randomize(data = NULL, unit_id = NULL, seed = NULL,
  intervention_type = "percentage", intervention_quantity = 50,
  block_id = NULL, baseline_vars = NULL)
}
\arguments{
\item{data}{CSV data file input}

\item{seed}{optional integer vector of length 1 containing a seed value to ensure identical randomization results across more than one run}

\item{block_id}{character vector of length 1 indicating the name of a variable indicating blocks within which randomization should occur, instead of across the overall data set}

\item{baseline_vars}{optional character vector of numeric variables for which to produce a balance plot between the treatment and comparison groups after completing random assignment}

\item{p}{optional numeric vector of length 1 containing a value greater than 0 and less than 1, indicating the probability of a single unit of analysis being assigned to the treatment group. Default value is 0.5.}
}
\value{
A list containing an error message if applicable, and if no errors are found, a file name containing the randomization results for download, and if baseline_vars are specified, the base64-encoded balance plot
}
\description{
Title
}
\examples{

}

