% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize_block}
\alias{randomize_block}
\title{Title}
\usage{
randomize_block(block_data, unit_id, intervention_type = "percentage",
  intervention_quantity = 50, baseline_vars = NULL)
}
\arguments{
\item{unit_id}{}
}
\description{
Title
}
\examples{

}
