% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope_probabilities.R
\name{rope_probabilities}
\alias{rope_probabilities}
\title{rope_probabilities}
\usage{
rope_probabilities(model, parameter, threshold)
}
\arguments{
\item{model}{The result of a call to easybayesian::stanlm}

\item{threshold}{A numeric cutoff value specified by the user.}
}
\value{
A list of three probabilities:
\code{less_than} is the probability that the parameter is less than the negative absolute value of the threshold.
\code{equal} is the probability that the parameter is within +/- absolute value of the threshold.
\code{greater_than} is the probability that the parameter is greater than the absolute value of the threshold.
}
\description{
rope_probabilities
}
\examples{

}
